fixed build - test case failed in hex encoding of aggpuubkey
