+++ src/test/xfieldhistory_tests.cpp
@@ -15,9 +15,9 @@
 struct XFieldHistorySetup : public TestingSetup {
     XFieldHistorySetup() : TestingSetup(TAPYRUS_MODES::DEV) {
         CXFieldHistory history;
-        history.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(std::vector<unsigned char>(ValidPubKeyStrings[10].begin(), ValidPubKeyStrings[10].end())), 20, uint256()));
-        history.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(std::vector<unsigned char>(ValidPubKeyStrings[11].begin(), ValidPubKeyStrings[11].end())), 40, uint256()));
-        history.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(std::vector<unsigned char>(ValidPubKeyStrings[12].begin(), ValidPubKeyStrings[12].end())), 60, uint256()));
+        history.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(CPubKey(ParseHex(ValidPubKeyStrings[10]))), 20, uint256()));
+        history.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(CPubKey(ParseHex(ValidPubKeyStrings[11]))), 40, uint256()));
+        history.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(CPubKey(ParseHex(ValidPubKeyStrings[12]))), 60, uint256()));
 
         history.Add(TAPYRUS_XFIELDTYPES::MAXBLOCKSIZE, XFieldChange(4000000, 30, uint256()));
         history.Add(TAPYRUS_XFIELDTYPES::MAXBLOCKSIZE, XFieldChange(8000000, 50, uint256()));
@@ -65,8 +65,8 @@ BOOST_AUTO_TEST_CASE(xfieldhistory_size_and_temp)
 BOOST_AUTO_TEST_CASE(xfieldhistory_add)
 {
     CXFieldHistory history1;
-    history1.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(std::vector<unsigned char>(ValidPubKeyStrings[13].begin(), ValidPubKeyStrings[13].end())), 70, uint256()));
-    history1.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(std::vector<unsigned char>(ValidPubKeyStrings[14].begin(), ValidPubKeyStrings[14].end())), 80, uint256()));
+    history1.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(CPubKey(ParseHex(ValidPubKeyStrings[13]))), 70, uint256()));
+    history1.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(CPubKey(ParseHex(ValidPubKeyStrings[14]))), 80, uint256()));
 
     //verify that the new history2 object uses the same map as history1
     CXFieldHistory history2;
@@ -76,14 +76,14 @@ BOOST_AUTO_TEST_CASE(xfieldhistory_add)
 
     //verify that the tempHistory object is not affected by changes to history1(hisrory2)
     CTempXFieldHistory tempHistory;
-    history2.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(std::vector<unsigned char>(ValidPubKeyStrings[15].begin(), ValidPubKeyStrings[15].end())), 80, uint256()));
-    history2.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(std::vector<unsigned char>(ValidPubKeyStrings[0].begin(), ValidPubKeyStrings[0].end())), 90, uint256()));
+    history2.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(CPubKey(ParseHex(ValidPubKeyStrings[15]))), 80, uint256()));
+    history2.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(CPubKey(ParseHex(ValidPubKeyStrings[0]))), 90, uint256()));
 
     BOOST_CHECK(tempHistory[TAPYRUS_XFIELDTYPES::AGGPUBKEY].size() != history2[TAPYRUS_XFIELDTYPES::AGGPUBKEY].size());
     BOOST_CHECK(tempHistory.Get(TAPYRUS_XFIELDTYPES::AGGPUBKEY, 91).height != history1.Get(TAPYRUS_XFIELDTYPES::AGGPUBKEY, 91).height);
 
     //verify that changes in tempHistory object are not reflected in history1(hisrory2)
-    tempHistory.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(std::vector<unsigned char>(ValidPubKeyStrings[1].begin(), ValidPubKeyStrings[1].end())), 100, uint256()));
+    tempHistory.Add(TAPYRUS_XFIELDTYPES::AGGPUBKEY, XFieldChange(XFieldAggPubKey(CPubKey(ParseHex(ValidPubKeyStrings[1]))), 100, uint256()));
     BOOST_CHECK(tempHistory.Get(TAPYRUS_XFIELDTYPES::AGGPUBKEY, 101).height != history1.Get(TAPYRUS_XFIELDTYPES::AGGPUBKEY, 101).height);
 
     tempHistory.Add(TAPYRUS_XFIELDTYPES::MAXBLOCKSIZE, XFieldChange(4000000, 91, uint256()));
@@ -100,15 +100,24 @@ BOOST_AUTO_TEST_CASE(XFieldChange_serialize)
     BOOST_CHECK_EQUAL(HexStr(stream.begin(), stream.end()), std::string("21025700236c2890233592fcef262f4520d22af9160e3d9705855140eb2aa06c35d300000000" + HexStr(FederationParams().GenesisBlock().GetHash())));
 
     stream.clear();
-    stream << history1.Get(TAPYRUS_XFIELDTYPES::AGGPUBKEY, 91);
+    history1.Get(TAPYRUS_XFIELDTYPES::AGGPUBKEY, 91).Serialize(stream);
 
-    BOOST_CHECK_EQUAL(HexStr(stream.begin(), stream.end()), "2103af80b90d25145da28c583359beb47b21796b2fe1a23c1511e443e7a64dfdb27d000000000000000000000000000000000000000000000000000000000000000000000000");
+    BOOST_CHECK_EQUAL(HexStr(stream.begin(), stream.end()), "2103af80b90d25145da28c583359beb47b21796b2fe1a23c1511e443e7a64dfdb27d5a0000000000000000000000000000000000000000000000000000000000000000000000");
+
+    stream.clear();
+    stream << history1.getXFieldHistoryMap().find(TAPYRUS_XFIELDTYPES::AGGPUBKEY)->second;
+
+    BOOST_CHECK_EQUAL(HexStr(stream.begin(), stream.end()), std::string("0821025700236c2890233592fcef262f4520d22af9160e3d9705855140eb2aa06c35d300000000" + HexStr(FederationParams().GenesisBlock().GetHash()) + "2103831a69b8009833ab5b0326012eaf489bfea35a7321b1ca15b11d88131423fafc1400000000000000000000000000000000000000000000000000000000000000000000002102bf2027c8455800c7626542219e6208b5fe787483689f1391d6d443ec85673ecf2800000000000000000000000000000000000000000000000000000000000000000000002103b44f1cfcf46aba8bc98e2fd39f137cc43d98ab7792e4848b09c06198b042ca8b3c00000000000000000000000000000000000000000000000000000000000000000000002102b9a609d6bec0fdc9ba690986013cf7bbd13c54ffc25e6cf30916b4732c4a952a4600000000000000000000000000000000000000000000000000000000000000000000002102e78cafe033b22bda5d7d1c8e82ee932930bf12e08489bc19769cbec765568be95000000000000000000000000000000000000000000000000000000000000000000000002102473757a955a23f75379820f3071abf5b3343b78eb54e52373d06259ffa6c550b5000000000000000000000000000000000000000000000000000000000000000000000002103af80b90d25145da28c583359beb47b21796b2fe1a23c1511e443e7a64dfdb27d5a0000000000000000000000000000000000000000000000000000000000000000000000"));
 
     stream.clear();
     stream << history1[TAPYRUS_XFIELDTYPES::MAXBLOCKSIZE][0];
     BOOST_CHECK_EQUAL(HexStr(stream.begin(), stream.end()),
     std::string("40420f0000000000" + HexStr(FederationParams().GenesisBlock().GetHash())));
 
+    stream.clear();
+    stream << history1.getXFieldHistoryMap().find(TAPYRUS_XFIELDTYPES::MAXBLOCKSIZE)->second;
+
+    BOOST_CHECK_EQUAL(HexStr(stream.begin(), stream.end()), std::string("0440420f0000000000" + HexStr(FederationParams().GenesisBlock().GetHash()) + "00093d001e000000000000000000000000000000000000000000000000000000000000000000000000127a003200000000000000000000000000000000000000000000000000000000000000000000000024f400460000000000000000000000000000000000000000000000000000000000000000000000"));
 }
 
 BOOST_AUTO_TEST_CASE(xfieldhistory_deserialize)
